improve how certificates are handledin the platform
