+++ lib/client.rb
@@ -3,4 +3,6 @@
 ###
 require_relative 'client/social'
 require_relative 'client/search'
+require_relative 'client/self_signed_certificate'
+require_relative 'client/ssl_socket'
 require_relative 'client/chrome_browser'
