+++ lib/client/self_signed_certificate.rb
@@ -1,0 +1,36 @@
+module Intrigue
+  class SelfSignedCertificate
+  
+    def generate
+      require 'openssl'
+  
+      key = OpenSSL::PKey::RSA.new(2048)
+      public_key = key.public_key
+  
+      subject = "/C=BE/O=Test/OU=Test/CN=Test"
+  
+      cert = OpenSSL::X509::Certificate.new
+      cert.subject = cert.issuer = OpenSSL::X509::Name.parse(subject)
+      cert.not_before = Time.now
+      cert.not_after = Time.now + 365 * 24 * 60 * 60
+      cert.public_key = public_key
+      cert.serial = 0x0
+      cert.version = 2
+  
+      ef = OpenSSL::X509::ExtensionFactory.new
+      ef.subject_certificate = cert
+      ef.issuer_certificate = cert
+      cert.extensions = [
+        ef.create_extension("basicConstraints","CA:TRUE", true),
+        ef.create_extension("subjectKeyIdentifier", "hash"),
+        # ef.create_extension("keyUsage", "cRLSign,keyCertSign", true),
+      ]
+      cert.add_extension ef.create_extension("authorityKeyIdentifier",
+                                             "keyid:always,issuer:always")
+  
+      cert.sign key, OpenSSL::Digest::SHA256.new
+    {"cert" => cert.to_pem, "key" => key.to_pem }
+    end
+  
+  end
+  end
