+++ lib/tasks/helpers/web.rb
@@ -133,210 +133,6 @@ module Task
     matching_urls.size.times.map {|x| matching_urls.pop }
   end
 
-  # See: https://raw.githubusercontent.com/zendesk/ruby-kafka/master/lib/kafka/ssl_socket_with_timeout.rb
-  def connect_ssl_socket(hostname, port, timeout=10, max_attempts=1)
-    # Create a socket and connect
-    # https://apidock.com/ruby/Net/HTTP/connect
-    attempts=0
-
-    begin
-
-      # keep track of how many times we've tried
-      attempts +=1
-
-      ssl_context = OpenSSL::SSL::SSLContext.new
-      #ssl_context.min_version = OpenSSL::SSL::SSL2_VERSION
-      #ssl_context.ssl_version = :SSLv23
-
-      # possible min versions:
-      # OpenSSL::SSL::SSL2_VERSION
-      # OpenSSL::SSL::SSL3_VERSION
-      # OpenSSL::SSL::TLS1_1_VERSION
-      # OpenSSL::SSL::TLS1_2_VERSION
-      # OpenSSL::SSL::TLS1_VERSION
-
-      # Open a tcp socket
-      addr = Socket.getaddrinfo(hostname, nil)
-      sockaddr = Socket.pack_sockaddr_in(port, addr[0][3])
-
-      tcp_socket = Socket.new(Socket.const_get(addr[0][0]), Socket::SOCK_STREAM, 0)
-      tcp_socket.setsockopt(Socket::IPPROTO_TCP, Socket::TCP_NODELAY, 1)
-
-      # first initiate the TCP socket
-      begin
-
-        # Initiate the socket connection in the background. If it doesn't fail
-        # immediately it will raise an IO::WaitWritable (Errno::EINPROGRESS)
-        # indicating the connection is in progress.
-        tcp_socket.connect_nonblock(sockaddr)
-        
-      rescue IO::WaitWritable
-        
-        # select will block until the socket is writable or the timeout
-        # is exceeded, whichever comes first.
-        unless _select_with_timeout(tcp_socket, :connect_write, timeout)
-          # select returns nil when the socket is not ready before timeout
-          # seconds have elapsed
-          tcp_socket.close
-          raise Errno::ETIMEDOUT
-        end
-
-        begin
-          # Verify there is now a good connection.
-          tcp_socket.connect_nonblock(sockaddr)
-        rescue Errno::EISCONN
-          # The socket is connected, we're good!
-        end
-
-      end
-
-      # once that's connected, we can start initiating the ssl socket
-      ssl_socket = OpenSSL::SSL::SSLSocket.new(tcp_socket, ssl_context)
-      ssl_socket.hostname = hostname # Required for SNI (cloudflare)
-
-      begin
-        # Initiate the socket connection in the background. If it doesn't fail
-        # immediately it will raise an IO::WaitWritable (Errno::EINPROGRESS)
-        # indicating the connection is in progress.
-        
-        # Unlike waiting for a tcp socket to connect, you can't time out ssl socket
-        # connections during the connect phase properly, because IO.select only partially works.
-        # Instead, you have to retry.
-
-        ssl_socket.connect_nonblock
-
-      rescue Errno::EAGAIN, Errno::EWOULDBLOCK, IO::WaitReadable
-        
-        if _select_with_timeout(ssl_socket, :connect_read, timeout)
-          _log "retrying... attempt: #{attempts}/#{max_attempts}"
-          retry unless attempts >= max_attempts
-        else
-          ssl_socket.close
-          tcp_socket.close
-          raise Errno::ETIMEDOUT
-        end
-
-      rescue IO::WaitWritable
-        
-        if _select_with_timeout(ssl_socket, :connect_write, timeout)
-          _log "retrying... attempt: #{attempts}/#{max_attempts}"
-          retry unless attempts >= max_attempts
-        else
-          ssl_socket.close
-          tcp_socket.close
-          raise Errno::ETIMEDOUT
-        end
-
-      end
-
-    rescue OpenSSL::SSL::SSLError => e
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}: #{e}"
-      _log "retrying... attempt: #{attempts}/#{max_attempts}"
-      retry unless attempts >= max_attempts
-    rescue SocketError => e
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::EINVAL => e
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::EMFILE => e
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::EAFNOSUPPORT => e
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::EPIPE => e
-      _log_error "Error requesting cert, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::ECONNRESET => e
-      _log_error "Error requesting cert, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::ECONNREFUSED => e
-      _log_error "Error requesting cert - refused, skipping: #{hostname} #{port}: #{e}"
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}"
-      _log "retrying... attempt: #{attempts}/#{max_attempts}"
-      retry unless attempts >= max_attempts
-    rescue Errno::ENETUNREACH
-      # unable to connect
-      _log_error "Error requesting cert, skipping: #{hostname} #{port}: #{e}"
-    rescue Errno::ETIMEDOUT => e
-      _log_error "Error requesting cert - timed out, timeout: #{hostname} #{port}: #{e}"
-      _log_error "Error requesting resource, skipping: #{hostname} #{port}"
-      _log "retrying... attempt: #{attempts}/#{max_attempts}"
-      retry unless attempts >= max_attempts
-    rescue Errno::EHOSTUNREACH => e
-      _log_error "Error requesting cert, skipping: #{hostname} #{port}: #{e}"
-    ensure
-      attempts +=1
-    end
-
-    # fail if we can't connect
-    unless ssl_socket
-      _log_error "Unable to connect!!"
-      return nil
-    end
-
-  ssl_socket
-  end
-
-  def _select_with_timeout(socket, type, timeout)
-    case type
-    when :connect_read
-      IO.select([socket], nil, nil, timeout)
-    when :connect_write
-      IO.select(nil, [socket], nil, timeout)
-    when :read
-      IO.select([socket], nil, nil, timeout)
-    when :write
-      IO.select(nil, [socket], nil, timeout)
-    end
-  end
-
-  def parse_names_from_cert(cert, skip_suspicious=true)
-
-    # default to empty alt_names
-    alt_names = []
-
-    # Check the subjectAltName property, and if we have names, here, parse them.
-    cert.extensions.each do |ext|
-      if "#{ext.oid}".match(/subjectAltName/)
-
-        alt_names = ext.value.split(",").collect do |x|
-          "#{x}".gsub(/DNS:/,"").gsub("IP Address:","").strip.gsub("*.","")
-        end
-        _log "Got cert's alt names: #{alt_names.inspect}"
-
-        tlds = []
-
-        # Iterate through, looking for trouble
-        alt_names.each do |alt_name|
-
-          # collect all top-level domains
-          tlds << alt_name.split(".").last(2).join(".")
-
-          universal_cert_domains = get_universal_cert_domains
-
-          universal_cert_domains.each do |cert_domain|
-            if alt_name.match(/#{cert_domain}$/) 
-              _log "This is a universal #{cert_domain} certificate, returning empty list"
-              return []
-            end
-          end
-
-        end
-
-        if skip_suspicious
-          # Generically try to find certs that aren't useful to us
-          suspicious_count = 20
-          # Check to see if we have over suspicious_count top level domains in this cert
-          if tlds.uniq.count >= suspicious_count
-            # and then check to make sure none of the domains are greate than a quarter
-            _log "This looks suspiciously like a third party cert... over #{suspicious_count} unique TLDs: #{tlds.uniq.count}"
-            _log "Total Unique Domains: #{alt_names.uniq.count}"
-            _log "Returning empty list"
-            return []
-          end
-        end
-      end
-    end
-
-  alt_names
-  end
-
 
   #
   # Download a file locally. Useful for situations where we need to parse the file
